Make RTC::Transport virtual and create a RTC::WebRtcTransport descendant class.
