+++ TODO_V2.md
@@ -1,12 +1,8 @@
 # TODO in mediasoup v2 (server-side)
 
-* Check `Destroy()` methods. Couldn't just the destructor be used?
-
 * Instead of using `static_cast<size_t>(number)` use `size_t{ number }`. NOTE: Investigate it.
 
-* Remove `producer.on('rtprawpacket')` and, instead, create a special `RtpConsumer` or something like that.
-
-* Set a proper value for Producer::FullFrameRequestBlockTimeout (currently 1 second).
+* Set a proper value for `Producer::FullFrameRequestBlockTimeout` (currently 1 second).
 
 * Check `appData` everywhere.
 
@@ -16,4 +12,4 @@
 
 * JS: Remove tons of useless debugs (for example in `Peer._createProducer()`).
 
-* Producer JS API to request fullframe.
+* Update `compile_commands_template.json`.
