+++ lib/Peer.js
@@ -700,10 +700,10 @@ class Peer extends EnhancedEventEmitter
 			transportId : id || utils.randomNumber()
 		};
 
-		return this._channel.request('router.createTransport', internal, options)
+		return this._channel.request('router.createWebRtcTransport', internal, options)
 			.then((data) =>
 			{
-				logger.debug('"router.createTransport" request succeeded');
+				logger.debug('"router.createWebRtcTransport" request succeeded');
 
 				data = Object.assign(data, { direction });
 
@@ -727,7 +727,8 @@ class Peer extends EnhancedEventEmitter
 			})
 			.catch((error) =>
 			{
-				logger.error('"router.createTransport" request failed: %s', String(error));
+				logger.error(
+					'"router.createWebRtcTransport" request failed: %s', String(error));
 
 				throw error;
 			});
