+++ worker/include/RTC/WebRtcTransport.hpp
@@ -1,0 +1,141 @@
+#ifndef MS_RTC_WEBRTC_TRANSPORT_HPP
+#define MS_RTC_WEBRTC_TRANSPORT_HPP
+
+#include "common.hpp"
+#include "Channel/Notifier.hpp"
+#include "RTC/DtlsTransport.hpp"
+#include "RTC/IceCandidate.hpp"
+#include "RTC/IceServer.hpp"
+#include "RTC/RemoteBitrateEstimator/RemoteBitrateEstimatorAbsSendTime.hpp"
+#include "RTC/SrtpSession.hpp"
+#include "RTC/StunMessage.hpp"
+#include "RTC/TcpConnection.hpp"
+#include "RTC/TcpServer.hpp"
+#include "RTC/Transport.hpp"
+#include "RTC/TransportTuple.hpp"
+#include "RTC/UdpSocket.hpp"
+#include <json/json.h>
+#include <string>
+#include <vector>
+
+namespace RTC
+{
+	class WebRtcTransport : public RTC::Transport,
+	                        public RTC::UdpSocket::Listener,
+	                        public RTC::TcpServer::Listener,
+	                        public RTC::TcpConnection::Listener,
+	                        public RTC::IceServer::Listener,
+	                        public RTC::DtlsTransport::Listener,
+	                        public RTC::RemoteBitrateEstimator::Listener
+	{
+	public:
+		struct TransportOptions
+		{
+			bool udp{ true };
+			bool tcp{ true };
+			bool preferIPv4{ true };
+			bool preferIPv6{ true };
+			bool preferUdp{ true };
+			bool preferTcp{ true };
+		};
+
+	public:
+		WebRtcTransport(
+		  RTC::Transport::Listener* listener,
+		  Channel::Notifier* notifier,
+		  uint32_t transportId,
+		  TransportOptions& options);
+
+	private:
+		~WebRtcTransport();
+
+	public:
+		Json::Value ToJson() const override;
+		RTC::DtlsTransport::Role SetRemoteDtlsParameters(
+		  RTC::DtlsTransport::Fingerprint& fingerprint, RTC::DtlsTransport::Role role);
+		void SetMaxBitrate(uint32_t bitrate);
+		void ChangeUfragPwd(std::string& usernameFragment, std::string& password);
+		void SendRtpPacket(RTC::RtpPacket* packet) override;
+		void SendRtcpPacket(RTC::RTCP::Packet* packet) override;
+
+	private:
+		bool IsConnected() const override;
+		void MayRunDtlsTransport();
+		void SendRtcpCompoundPacket(RTC::RTCP::CompoundPacket* packet) override;
+
+		/* Private methods to unify UDP and TCP behavior. */
+	private:
+		void OnPacketRecv(RTC::TransportTuple* tuple, const uint8_t* data, size_t len);
+		void OnStunDataRecv(RTC::TransportTuple* tuple, const uint8_t* data, size_t len);
+		void OnDtlsDataRecv(const RTC::TransportTuple* tuple, const uint8_t* data, size_t len);
+		void OnRtpDataRecv(RTC::TransportTuple* tuple, const uint8_t* data, size_t len);
+		void OnRtcpDataRecv(RTC::TransportTuple* tuple, const uint8_t* data, size_t len);
+
+		/* Pure virtual methods inherited from RTC::UdpSocket::Listener. */
+	public:
+		void OnPacketRecv(
+		  RTC::UdpSocket* socket, const uint8_t* data, size_t len, const struct sockaddr* remoteAddr) override;
+
+		/* Pure virtual methods inherited from RTC::TcpServer::Listener. */
+	public:
+		void OnRtcTcpConnectionClosed(
+		  RTC::TcpServer* tcpServer, RTC::TcpConnection* connection, bool isClosedByPeer) override;
+
+		/* Pure virtual methods inherited from RTC::TcpConnection::Listener. */
+	public:
+		void OnPacketRecv(RTC::TcpConnection* connection, const uint8_t* data, size_t len) override;
+
+		/* Pure virtual methods inherited from RTC::IceServer::Listener. */
+	public:
+		void OnOutgoingStunMessage(
+		  const RTC::IceServer* iceServer, const RTC::StunMessage* msg, RTC::TransportTuple* tuple) override;
+		void OnIceSelectedTuple(const RTC::IceServer* iceServer, RTC::TransportTuple* tuple) override;
+		void OnIceConnected(const RTC::IceServer* iceServer) override;
+		void OnIceCompleted(const RTC::IceServer* iceServer) override;
+		void OnIceDisconnected(const RTC::IceServer* iceServer) override;
+
+		/* Pure virtual methods inherited from RTC::DtlsTransport::Listener. */
+	public:
+		void OnDtlsConnecting(const RTC::DtlsTransport* dtlsTransport) override;
+		void OnDtlsConnected(
+		  const RTC::DtlsTransport* dtlsTransport,
+		  RTC::SrtpSession::Profile srtpProfile,
+		  uint8_t* srtpLocalKey,
+		  size_t srtpLocalKeyLen,
+		  uint8_t* srtpRemoteKey,
+		  size_t srtpRemoteKeyLen,
+		  std::string& remoteCert) override;
+		void OnDtlsFailed(const RTC::DtlsTransport* dtlsTransport) override;
+		void OnDtlsClosed(const RTC::DtlsTransport* dtlsTransport) override;
+		void OnOutgoingDtlsData(
+		  const RTC::DtlsTransport* dtlsTransport, const uint8_t* data, size_t len) override;
+		void OnDtlsApplicationData(
+		  const RTC::DtlsTransport* dtlsTransport, const uint8_t* data, size_t len) override;
+
+		/* Pure virtual methods inherited from RTC::RemoteBitrateEstimator::Listener. */
+	public:
+		void OnRemoteBitrateEstimatorValue(const std::vector<uint32_t>& ssrcs, uint32_t bitrate) override;
+
+	private:
+		// Allocated by this.
+		RTC::IceServer* iceServer{ nullptr };
+		std::vector<RTC::UdpSocket*> udpSockets;
+		std::vector<RTC::TcpServer*> tcpServers;
+		RTC::DtlsTransport* dtlsTransport{ nullptr };
+		RTC::SrtpSession* srtpRecvSession{ nullptr };
+		RTC::SrtpSession* srtpSendSession{ nullptr };
+		// Others (ICE).
+		std::vector<IceCandidate> iceLocalCandidates;
+		RTC::TransportTuple* selectedTuple{ nullptr };
+		// Others (DTLS).
+		bool hasRemoteDtlsParameters{ false };
+		RTC::DtlsTransport::Role dtlsLocalRole{ RTC::DtlsTransport::Role::AUTO };
+		// Others (REMB and bitrate stuff).
+		std::unique_ptr<RTC::RemoteBitrateEstimatorAbsSendTime> remoteBitrateEstimator;
+		uint32_t maxBitrate{ 0 };
+		uint32_t effectiveMaxBitrate{ 0 };
+		uint64_t lastEffectiveMaxBitrateAt{ 0 };
+	};
+} // namespace RTC
+
+#endif
