+++ worker/mediasoup-worker.gyp
@@ -41,6 +41,7 @@
       'src/RTC/Transport.cpp',
       'src/RTC/TransportTuple.cpp',
       'src/RTC/UdpSocket.cpp',
+      'src/RTC/WebRtcTransport.cpp',
       'src/RTC/Codecs/Codecs.cpp',
       'src/RTC/Codecs/VP8.cpp',
       'src/RTC/RtpDictionaries/Media.cpp',
@@ -131,6 +132,7 @@
       'include/RTC/Transport.hpp',
       'include/RTC/TransportTuple.hpp',
       'include/RTC/UdpSocket.hpp',
+      'include/RTC/WebRtcTransport.hpp',
       'include/RTC/Codecs/Codecs.hpp',
       'include/RTC/Codecs/VP8.hpp',
       'include/RTC/RTCP/Packet.hpp',
